chore: set mock types to mirror runtime ones

+++ primitives/src/bridges/chain_circuit/mod.rs
@@ -109,7 +109,7 @@ pub mod time_units {
 }
 
 /// Block number type used in Circuit.
-pub type BlockNumber = u64;
+pub type BlockNumber = u32;
 
 /// Hash type used in Circuit.
 pub type Hash = <BlakeTwo256 as HasherT>::Out;
